requirements.txt
Flask==2.3.3
openai-whisper==20240930
Flask-Cors==4.0.1